+++ src/gui/viewModels/ChartViewModel.java
@@ -80,11 +80,11 @@ public class ChartViewModel{
 
         //Technician is unique by its UserID
         ticketFacade.getAllTechnicians()
-                .forEach(tech -> mapTechniciansResolved.put(tech.getUserId(), 0));
+                .forEach(tech -> mapTechniciansResolved.put((long) tech.getUserId(), 0));
 
         ticketFacade.giveActemiumTicketsResolved()
                 .forEach(ticket -> ticket.giveTechnicians()
-                        .forEach(tech -> mapTechniciansResolved.put(tech.getUserId(), mapTechniciansResolved.get(tech.getUserId()) + 1)));
+                        .forEach(tech -> mapTechniciansResolved.put((long) tech.getUserId(), mapTechniciansResolved.get(tech.getUserId()) + 1)));
 
         return mapTechniciansResolved;
     }
@@ -94,11 +94,11 @@ public class ChartViewModel{
 
         //Technician is unique by its UserID
         ticketFacade.getAllTechnicians()
-                .forEach(tech -> mapTechniciansOutstanding.put(tech.getUserId(), 0));
+                .forEach(tech -> mapTechniciansOutstanding.put((long) tech.getUserId(), 0));
 
         ticketFacade.giveActemiumTicketsOutstanding()
                 .forEach(ticket -> ticket.giveTechnicians()
-                        .forEach(tech -> mapTechniciansOutstanding.put(tech.getUserId(), mapTechniciansOutstanding.get(tech.getUserId()) + 1)));
+                        .forEach(tech -> mapTechniciansOutstanding.put((long) tech.getUserId(), mapTechniciansOutstanding.get(tech.getUserId()) + 1)));
 
         return mapTechniciansOutstanding;
     }
